spark/requirements-spark.txt
redis==6.2.0
psycopg2-binary==2.9.9